# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kaynatt API",
        "description": "Mon API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/createur_tontine/gererCycle/{tontine}": {
            "post": {
                "tags": [
                    "CreateurTontine"
                ],
                "summary": "Gérer les cycles d'une tontine",
                "description": "Crée et liste les cycles d'une tontine en fonction de sa période.",
                "operationId": "gestionCycle",
                "parameters": [
                    {
                        "name": "tontine",
                        "in": "path",
                        "description": "ID de la tontine",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès - Les cycles de la tontine ont été gérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "les cycles du tontine"
                                        },
                                        "cycles": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "dates": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "2024-02-10"
                                            }
                                        },
                                        "nombre_participants": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé - L'utilisateur n'est pas le créateur de la tontine"
                    },
                    "400": {
                        "description": "Requête incorrecte - La tontine n'est pas dans un état accepté ou a déjà des cycles"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/auth/ParticiperTontine": {
            "post": {
                "tags": [
                    "ParticipationTontines"
                ],
                "summary": "Demander à participer à une tontine",
                "description": "Demander à participer à une tontine.",
                "operationId": "feaa79a2163a7d71a3043a20e8e04a94",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tontine_id",
                                    "date"
                                ],
                                "properties": {
                                    "tontine_id": {
                                        "description": "ID de la tontine à laquelle participer",
                                        "type": "integer"
                                    },
                                    "date": {
                                        "description": "Date de participation",
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Demande de participation à la tontine effectuée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "Votre demande de participation sera approuvée ou déclinée par le créateur de cette tontine. Vous recevrez une notification bientôt."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Vous avez déjà participé à cette tontine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "Vous avez déjà participé à cette tontine."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/createur_tontine/AcceptedParticipationUser/{participationTontines}": {
            "post": {
                "tags": [
                    "CreateurTontine"
                ],
                "summary": "Accepter la participation d'un utilisateur à une tontine",
                "description": "Accepter la participation d'un utilisateur à une tontine par le créateur de la tontine.",
                "operationId": "accepteParticipation",
                "parameters": [
                    {
                        "name": "participationTontines",
                        "in": "path",
                        "description": "L'ID de la participation à la tontine à accepter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès - La demande de participation à la tontine a été acceptée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Indique si l'opération a réussi ou non",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Message indiquant le résultat de l'opération",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé - L'utilisateur n'est pas le créateur de la tontine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Indique si l'opération a réussi ou non",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Message d'erreur indiquant l'accès non autorisé",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}